+++ lib/src/spaced_row.dart
@@ -1,24 +1,37 @@
 part of flutter_widgetz;
 
 /// {@template flutter_widgetz.SpacedRow}
-/// Spaces [children] with the given spacing.
+/// Spaces a row of [children] with the given spacing.
 /// {@endtemplate}
 class SpacedRow extends StatelessWidget {
   /// {@macro flutter_widgetz.SpacedRow}
   const SpacedRow({
     Key? key,
     required this.children,
-    this.title,
-    this.spacing = 20.0,
     this.crossAxisAlignment = CrossAxisAlignment.start,
     this.mainAxisAlignment = MainAxisAlignment.spaceEvenly,
+    this.spacing = 20.0,
+    this.title,
+    this.titleStyle,
   }) : super(key: key);
 
+  /// Creates a horizontal array of children.
   final List<Widget> children;
-  final String? title;
-  final double spacing;
+
+  /// How the children should be placed along the cross axis.
   final CrossAxisAlignment crossAxisAlignment;
+
+  /// How the children should be placed along the main axis.
   final MainAxisAlignment mainAxisAlignment;
+
+  /// The spacing between the children.
+  final double spacing;
+
+  /// An optional title to display above this row.
+  final String? title;
+
+  /// The style to use for the title.
+  final TextStyle? titleStyle;
 
   @override
   Widget build(BuildContext context) {
@@ -31,26 +44,28 @@ class SpacedRow extends StatelessWidget {
         if (title != null)
           Text(
             title!,
-            style: Theme.of(context).textTheme.titleSmall,
+            style: titleStyle ?? Theme.of(context).textTheme.titleSmall,
           ),
         Row(
           crossAxisAlignment: crossAxisAlignment,
           mainAxisAlignment: mainAxisAlignment,
-          children: _formatChildren(),
+          children: _getChildren(),
         ),
       ],
     );
   }
 
-  List<Widget> _formatChildren() {
+  List<Widget> _getChildren() {
     final List<Widget> result = <Widget>[];
 
     for (int i = 0; i < children.length; i++) {
+      final Widget child = children[i];
+
       if (i != children.length - 1) {
-        result.add(Expanded(child: children[i]));
+        result.add(Expanded(child: child));
         result.add(SizedBox(width: spacing));
       } else {
-        result.add(Expanded(child: children[i]));
+        result.add(Expanded(child: child));
       }
     }