+++ lib/src/text_form_field.dart
@@ -8,12 +8,13 @@ class CustomTextFormField extends StatelessWidget {
   /// {@macro flutter_widgetz.CustomTextFormField}
   const CustomTextFormField({
     Key? key,
-    this.autoFillHint,
+    this.autofillHint,
     this.autofocus = false,
     this.controller,
     this.errorText,
     this.focusNode,
     this.helpText,
+    this.hintText,
     this.initialValue,
     this.inputFormatter,
     this.keyboardType,
@@ -21,41 +22,83 @@ class CustomTextFormField extends StatelessWidget {
     this.maxLines = 1,
     this.minLines,
     this.obscureText = false,
+    this.onChanged,
     this.prefixIcon,
     this.scrollPadding = const EdgeInsets.all(20.0),
     this.showError = false,
     this.suffixIcon,
     this.textCapitalization = TextCapitalization.none,
     this.textInputAction,
-    this.onChanged,
   }) : super(key: key);
 
-  final String? autoFillHint;
+  /// Determines what the input field expects.
+  final String? autofillHint;
+
+  /// Determines whether to auto focus this input.
   final bool autofocus;
+
+  /// Controls the text being edited.
   final TextEditingController? controller;
+
+  /// The text shown when there is an error.
   final String? errorText;
+
+  /// Used to obtain the keyboard focus and to handle keyboard events.
   final FocusNode? focusNode;
+
+  /// If given, shows a suffix icon and dialog with this string.
   final String? helpText;
+
+  /// Text that suggests what sort of input the field accepts.
+  final String? hintText;
+
+  /// The initial value of this input.
   final String? initialValue;
+
+  /// Provides as-you-type validation and formatting of the text being edited.
   final TextInputFormatter? inputFormatter;
+
+  /// The type of information for which to optimize the text input control.
   final TextInputType? keyboardType;
+
+  /// An optional label for this input.
   final String? labelText;
+
+  /// The maximum number of lines for this input.
   final int? maxLines;
+
+  /// The minimum number of lines for this input.
   final int? minLines;
+
+  /// Determines whether to obscure the text in this input.
   final bool obscureText;
-  final IconData? prefixIcon;
-  final EdgeInsets scrollPadding;
-  final bool showError;
-  final Widget? suffixIcon;
-  final TextCapitalization textCapitalization;
-  final TextInputAction? textInputAction;
+
+  /// Called when the value of the input changes.
   final ValueChanged<String>? onChanged;
+
+  /// An icon that appers before the editable part of the text field.
+  final IconData? prefixIcon;
+
+  /// Configures padding when the input scrolls into view.
+  final EdgeInsets scrollPadding;
+
+  /// Determines if an error should be displayed.
+  final bool showError;
+
+  /// An icon that appers after the editable part of the text field.
+  final Widget? suffixIcon;
+
+  /// Determines whether to show an uppercase or lowercase keyboard.
+  final TextCapitalization textCapitalization;
+
+  /// An action the user has requested the text input control to perform.
+  final TextInputAction? textInputAction;
 
   @override
   Widget build(BuildContext context) {
     return TextFormField(
       autofocus: autofocus,
-      autofillHints: autoFillHint != null ? <String>[autoFillHint!] : null,
+      autofillHints: autofillHint != null ? <String>[autofillHint!] : null,
       controller: controller,
       focusNode: focusNode,
       initialValue: initialValue,
@@ -70,6 +113,7 @@ class CustomTextFormField extends StatelessWidget {
       textInputAction: textInputAction,
       decoration: InputDecoration(
         errorText: showError ? errorText : null,
+        hintText: hintText,
         labelText: labelText,
         prefixIcon: prefixIcon != null ? Icon(prefixIcon) : null,
         suffixIcon: _getSuffixIcon(context),